+++ python/src/chauff_cmn/__init__.py
@@ -1,3 +1,3 @@
-from chauff_cmn import logging, models
+from chauff_cmn import logging, models, tracing
 
-__all__ = ["logging", "models"]
+__all__ = ["logging", "models", "tracing"]