+++ typescript/src/index.ts
@@ -1,5 +1,12 @@
 export * from "./models";
 export { logger, configure } from "./logging";
-export { TRACEPARENT_HEADER, resolveTraceId, runWithTraceId } from "./logging/_trace";
+export {
+  setupTracing,
+  shutdownTracing,
+  activeTraceContext,
+  extractTraceContext,
+  withTraceContext,
+  type Context,
+} from "./tracing";
 export { withRequestLogging } from "./logging/request";
 export { tracedFetch } from "./logging/fetch";